refactor : PathService 메소드 분리

+++ src/main/java/wooteco/subway/admin/domain/CriteriaType.java
@@ -23,7 +23,7 @@ public enum CriteriaType {
             .orElseThrow(AssertionError::new);
     }
 
-    public double get(LineStation it) {
+    public int get(LineStation it) {
         return this.expression.apply(it);
     }
 }
